railway_service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python
python-dotenv
PyJWT==2.8.0
waitress
